refactor: move some methods to new utils file
